2020.12.29 버그 수정

+++ ios/Classes/SensorController.swift
@@ -49,7 +49,7 @@ class SensorController: NSObject, CLLocationManagerDelegate {
             // 제한 시간 설정
             if let start = startDate{
                 let interval = Date().timeIntervalSince(start)
-                if interval >= (60*60*TIME_LIMIT) { // 제한시간
+                if interval >= Double(60 * 60 * TIME_LIMIT) { // 제한시간
                     locationManager.stopUpdatingLocation()
                     locationManager.stopMonitoringSignificantLocationChanges()
                 }
